+++ udata_front/theme/gouvfr/assets/js/components/Accordion/AccordionGroup.vue
@@ -4,8 +4,5 @@
   </div>
 </template>
 
-<script>
-import { defineComponent } from 'vue';
-
-export default defineComponent({});
+<script setup lang="ts">
 </script>