+++ udata_front/theme/gouvfr/datagouv-components/src/components/index.ts
@@ -1,6 +1,8 @@
 import { CopyButton } from "./CopyButton";
 import InformationPanel from "./InformationPanel";
 import OrganizationNameWithCertificate from "./Organization";
+import Pagination from "./Pagination";
 import { ResourceAccordion, Loader as ResourceAccordionLoader } from "./ResourceAccordion";
+import Well from "./Well";
 
-export { CopyButton, InformationPanel, OrganizationNameWithCertificate, ResourceAccordion, ResourceAccordionLoader };
+export { CopyButton, InformationPanel, OrganizationNameWithCertificate, Pagination, ResourceAccordion, ResourceAccordionLoader, Well };